add auth validation

+++ src/controllers/products.controller.js
@@ -1,4 +1,5 @@
 import { productsService } from "../repositories/index.js";
+import CustomError from "../services/errors/CustomError.js";
 
 export const getProducts = async (req, res) => {
   try {
@@ -23,9 +24,12 @@ export const getProduct = async (req, res) => {
 
 export const addProduct = async (req, res) => {
   try {
+    const { role, id } = req.user
     const product = req.body;
-    const result = await productsService.createProduct(product);
 
+    if(role === "premium") product.owner = id
+
+    const result = await productsService.createProduct(product);
     res.json({ status: "success", payload: result });
   } catch (error) {
     req.logger.error(error.toString());
@@ -37,6 +41,14 @@ export const updateProduct = async (req, res) => {
   try {
     const pid = req.params.pid;
     const product = await productsService.getProduct(pid);
+    const user = req.user;
+
+    if(user.role === "premium" && user.id !== product.owner) {
+      const error = "You can't modify a product owned by another user"
+      req.logger.error(error)
+      return res.status(403).json({status: "error", error})
+    }
+
     const updatedProduct = {
       ...product,
       ...req.body,
@@ -52,6 +64,15 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   try {
     const pid = req.params.pid;
+    const product = productsService.getProduct(pid);
+    const user = req.user;
+
+    if(user.role === "premium" && user.id !== product.owner) {
+      const error = "You can't modify a product owned by another user"
+      req.logger.error(error)
+      return res.status(403).json({status: "error", error})
+    }
+
     const result = await productsService.deleteProduct(pid);
 
     res.json({ status: "success", payload: result });
